ready to flatten options
